ebs for all workers
use an ebs volume for the jenkins home on all workers
(local SSD is no longer available on EC2 c4)

uniform naming/labelling for publish<->public workers

deal with multiple jdks by including java recipe last
(it updates alternatives)

include apt recipe to do apt-get update on debian

move to more recent ami for ubuntu utopic

+++ recipes/_worker-config-ebs.rb
@@ -7,36 +7,56 @@
 # All rights reserved - Do Not Redistribute
 #
 
-jenkinsHome="/home/jenkins"
-jenkinsUser="jenkins"
-jenkinsDev="/dev/sdj"
-jenkinsFs="ext4"
+node['ebs']['volumes'].each do |mountPoint, ebsConfig|
+  aws_ebs_volume ebsConfig['dev'] do
+    size    ebsConfig['size']
+    device  ebsConfig['dev']
 
-aws_ebs_volume jenkinsDev do
-  size 100
-  device jenkinsDev
-  volume_type "gp2"
-  availability_zone node[:ec2][:placement_availability_zone]
-  action [:create, :attach]
-end
-
-execute 'mkfs' do
-  command "mkfs -t #{jenkinsFs} #{jenkinsDev}"
-  not_if do
-    BlockDevice.wait_for(jenkinsDev)
-    system("blkid -s TYPE -o value #{jenkinsDev}")
+    volume_type "gp2"
+    availability_zone node[:ec2][:placement_availability_zone]
+    action [:create, :attach]
   end
-end
 
-directory jenkinsHome do
-  owner jenkinsUser
-  action :create
-  mode 0755
-end
+  case node["platform_family"]
+  when "windows"
+    force="" # set to "NOERR" to force partitioning&formatting the disk
+    diskpartScript=<<-EOX.gsub(/^    /, '')
+    select disk #{ebsConfig['disk']}
+    attributes disk clear readonly
+    online disk #{force}
+    create partition primary #{force}
+    select partition 1
+    format FS=#{ebsConfig['fstype']} quick
+    assign LETTER=#{mountPoint}
+    EOX
+    script "setupdisk" do
+      interpreter "diskpart"
+      flags "/s"
+      code diskpartScript
+      not_if { ::File.directory?("#{mountPoint}:\\") }
+    end
+  else
+    execute 'mkfs' do
+      command "mkfs -t #{ebsConfig['fstype']} #{ebsConfig['dev']}"
+      not_if do
+        BlockDevice.wait_for(ebsConfig['dev'])
+        system("blkid -s TYPE -o value #{ebsConfig['dev']}")
+      end
+    end
 
-mount jenkinsHome do
-  device jenkinsDev
-  fstype jenkinsFs
-  options 'noatime'
-  action [:mount, :enable]
+    directory mountPoint do
+      owner ebsConfig['user']
+      mode 0755
+
+      action :create
+    end
+
+    mount mountPoint do
+      device  ebsConfig['dev']
+      fstype  ebsConfig['fstype']
+      options ebsConfig['mountopts']
+
+      action [:mount, :enable]
+    end
+  end
 end
